refactor(search): keep the overlay seams internal (no public ABI leak)
The CI tarball-smoke gate failed on dist/index.d.mts leaking private
pi-tui / internal paths (src/transcript.d.ts and its type deps
icons/present/token-usage): the search-overlay seams
(SearchOverlayState / refreshedSearchState / steppedSearchOverlayState)
were exported from the runner's public entry and their signatures
reference TranscriptFolder and TranscriptSearchMatch.

The seams move to src/search-overlay.ts — an INTERNAL module the runner
and the headless tests import directly, so the published .d.mts no
longer drags the private transcript declaration in (verified: no
'src/transcript' region remains in dist/index.d.mts).

Verified: pnpm pack:release passes end to end (prepack build+typecheck+
bundle tests, postpack tarball smoke), full bundle suite green.
